+++ tests/algos/doe/test_doe_lib.py
@@ -18,9 +18,10 @@
 #    INITIAL AUTHORS - API and implementation and/or documentation
 #        :author: Francois Gallard
 #    OTHER AUTHORS   - MACROSCOPIC CHANGES
+from unittest import mock
 
 import pytest
-from numpy import array
+from numpy import arange, array
 
 from gemseo.algos.doe.doe_factory import DOEFactory
 from gemseo.algos.doe.doe_lib import DOELibrary
@@ -65,3 +66,30 @@ def test_phip_criteria():
     samples = array([[0.0, 0.0], [0.0, 2.0], [0.0, 3.0]])
     expected = sum([val ** (-power) for val in [2.0, 3.0, 1.0]]) ** (1.0 / power)
     assert DOELibrary.compute_phip_criteria(samples, power) == expected
+
+
+@pytest.fixture(scope="module")
+def variables_space():
+    """A mock design space."""
+    design_space = mock.Mock()
+    design_space.dimension = 2
+    design_space.unnormalize_vect = mock.Mock(return_value=arange(6).reshape((3, 2)))
+    return design_space
+
+
+def test_compute_doe_normalized(doe, variables_space):
+    """Check the computation of a normalized DOE in a variables space."""
+    doe.algo_name = "lhs"
+    points = doe.compute_doe(variables_space, 3, normalize=True)
+    assert points.shape == (3, 2)
+    assert points.max() <= 1.0
+    assert points.min() >= 0.0
+    variables_space.unnormalize_vect.assert_not_called()
+
+
+def test_compute_doe_nonnormalized(doe, variables_space):
+    """Check the computation of a non-normalized DOE in a variables space."""
+    doe.algo_name = "lhs"
+    points = doe.compute_doe(variables_space, 3)
+    assert points.shape == (3, 2)
+    variables_space.unnormalize_vect.assert_called_once()